Switch to new functional tests
